refactor: move Pipeline implementation into ml/pipeline.py; thin ml/__init__
ml/__init__.py held the entire ~300-line Pipeline + PipelineState. Extract them
into ml/pipeline.py and reduce __init__ to a facade (docstring + re-export
Pipeline, PipelineState, save_pickle, load_pickle). Matches the sources/outputs/
recipes/bridges convention. Public API (from myogestic.ml import Pipeline, ...)
unchanged.
